functions/google_signin/requirements.txt
requests
passlib==1.7.4
pydantic==1.10.12
loguru
bcrypt
python-jose[cryptography]
cryptography==38.0.3
google-auth
google-auth-oauthlib
